Quinze de junho, correção try catch

+++ QuinzeDeJunho/Calculadora.cs
@@ -6,31 +6,31 @@ namespace QuinzeDeJunho
 {
     abstract class Calculadora
     {
-        public double Valor1 { get; set; }
-        public double Valor2 { get; set; }
-        public double Valor3 { get; set; }
+        public decimal Valor1 { get; set; }
+        public decimal Valor2 { get; set; }
+        public decimal Valor3 { get; set; }
 
-        public Calculadora(double valor1, double valor2)
+        public Calculadora(decimal valor1, decimal valor2)
         {
             Valor1 = valor1;
             Valor2 = valor2;
         }
-        public double Soma()
+        public decimal Soma()
         {
             Valor3 = Valor1 + Valor2;
             return Valor1 + Valor2;
         }
-        public double Subtracao()
+        public decimal Subtracao()
         {
             Valor3 = Valor1 - Valor2;
             return Valor1 - Valor2;
         }
-        public double Divisao()
+        public decimal Divisao()
         {
             Valor3 = Valor1 / Valor2;
             return Valor1 / Valor2;
         }
-        public double Multiplicacao()
+        public decimal Multiplicacao()
         {
             Valor3 = Valor1 * Valor2;
             return Valor1 * Valor2;
